requirements.txt
fastapi
uvicorn
python-multipart
psutil
face_recognition
opencv-python
numpy
python-dotenv
requests